Add files via upload
This script will perform a virual PFGE using different restriction enzymes of user's choice ~(more enzymes can be added to the list). It uses .fasta files but not multifasta (such as contigs from the assembly). It works with fasta files generated by Prokka (.fna but you should rename them to .fasta). It will work also with file generated by other software as far they are fasta and not multifasta